Refactor data import panel
