+++ src/dashboard/dataImport/main.ts
@@ -1,0 +1,5 @@
+import '../globalStyles.css';
+import './dataImport.css';
+
+import './teamData';
+import './roundData';
