+++ src/types/importStatus.ts
@@ -1,0 +1,5 @@
+export enum ImportStatus {
+    Success,
+    Loading,
+    Failure
+}
